refactor MySQLResponseService in order to support more protocol
