+++ src/main/java/com/actiontech/dble/net/service/Service.java
@@ -11,7 +11,5 @@ public interface Service {
 
     void execute(ServiceTask task);
 
-    void consumerInternalData(ServiceTask task);
-
     void cleanup();
 }
